update input parameter to aom_wb_write_uvlc

+++ common/av1_config.c
@@ -531,7 +531,7 @@ int write_av1config(const Av1Config *config, size_t capacity,
   aom_wb_write_literal(&writer, config->seq_level_idx_0, 5);
   aom_wb_write_bit(&writer, config->seq_tier_0);
 #if CONFIG_CWG_E242_BITDEPTH
-  aom_wb_write_uvlc(&writer, config->bitdepth_idx);
+  aom_wb_write_uvlc(&writer, (uint32_t)config->bitdepth_idx);
 #else
   aom_wb_write_bit(&writer, config->high_bitdepth);
   aom_wb_write_bit(&writer, config->twelve_bit);
